Move everything to net.codingwell per codingwell/scala-guice#8
