+++ src/main/scala/net/codingwell/scalaguice/SetProvider.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.me.lings.scalaguice
+package net.codingwell.scalaguice
 
 import com.google.common.collect.ImmutableSet
 import com.google.inject._
