+++ src/main/scala/net/codingwell/scalaguice/binder/BindingProxies.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.me.lings.scalaguice
+package net.codingwell.scalaguice
 package binder
 
 import com.google.inject._
