+++ src/test/scala/net/codingwell/scalaguice/BindingExtensionsSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.me.lings.scalaguice
+package net.codingwell.scalaguice
 
 import org.scalatest.WordSpec
 import org.scalatest.matchers.ShouldMatchers
@@ -21,7 +21,7 @@ import org.scalatest.matchers.ShouldMatchers
 import com.google.inject._
 
 class BindingExtensionsSpec extends WordSpec with ShouldMatchers {
-  
+
   import BindingExtensions._
 
   def module(body: Binder => Unit) =  new Module {
@@ -29,7 +29,7 @@ class BindingExtensionsSpec extends WordSpec with ShouldMatchers {
   }
 
   "Binding extensions" should {
-  
+
     "allow binding source type using a type parameter" in {
       Guice createInjector module { binder =>
         binder.bindType[A].to(classOf[B])
@@ -48,27 +48,27 @@ class BindingExtensionsSpec extends WordSpec with ShouldMatchers {
       } getInstance(new Key[Gen[String]] {})
       inst.get should equal ("String")
     }
-    
+
     "allow binding between nested types" in {
       val inst = Guice createInjector module { binder =>
           binder.bindType[Outer.Gen[String]].toType[Outer.C]
       } getInstance(new Key[Outer.Gen[String]] {})
       inst.get should equal ("String")
     }
-    
-    "allow binding to provider using type parameter" in { 
+
+    "allow binding to provider using type parameter" in {
       val inst = Guice createInjector module { binder =>
           binder.bindType[Gen[String]].toProviderType[GenStringProvider]
       } getInstance(new Key[Gen[String]] {})
       inst.get should equal ("String")
     }
-    
-    "allow binding to provider of subtype using type parameter" in { 
+
+    "allow binding to provider of subtype using type parameter" in {
       val inst = Guice createInjector module { binder =>
           binder.bindType[Gen[String]].toProviderType[CProvider]
       } getInstance(new Key[Gen[String]] {})
       inst.get should equal ("String")
-    } 
+    }
 
   }
