+++ src/test/scala/net/codingwell/scalaguice/ScalaPrivateModuleSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.me.lings.scalaguice
+package net.codingwell.scalaguice
 
 import org.scalatest.WordSpec
 import org.scalatest.matchers.ShouldMatchers
